# Improve injection performance