--- a/pkg/kube/inject/template.go
+++ b/pkg/kube/inject/template.go
@@ -149,7 +149,6 @@ func fromJSON(j string) interface{} {
 		return "{}"
 	}
 
-	log.Warnf("%v", m)
 	return m
 }