Fixes https://github.com/istio/istio/issues/34799
Inspired by https://github.com/istio/istio/pull/34803 (we may still want that one as well)
```
$ cat /tmp/old
goos: linux
goarch: amd64
pkg: istio.io/istio/pkg/kube/inject
cpu: Intel(R) Core(TM) i7-8665U CPU @ 1.90GHz
       10        12302795235 ns/op       3494954690 B/op  7523207 allocs/op
       10        12152944856 ns/op       3495193698 B/op  7523328 allocs/op
       10        13032458981 ns/op       3495241792 B/op  7523395 allocs/op
PASS
ok      istio.io/istio/pkg/kube/inject  158.548s
$ cat /tmp/new
goos: linux
goarch: amd64
pkg: istio.io/istio/pkg/kube/inject
cpu: Intel(R) Core(TM) i7-8665U CPU @ 1.90GHz
      10         858556624 ns/op        351448878 B/op    585319 allocs/op
      10        1010176965 ns/op        351284525 B/op    585313 allocs/op
      10        1092338275 ns/op        351240620 B/op    585312 allocs/op
PASS
ok      istio.io/istio/pkg/kube/inject  36.286s
```

About 14x slower. Example comes from a real world pod.

**Please provide a description of this PR:**



**To help us figure out who should review this PR, please put an X in all the areas that this PR affects.**

- [ ] Configuration Infrastructure
- [ ] Docs
- [ ] Installation
- [ ] Networking
- [ ] Performance and Scalability
- [ ] Policies and Telemetry
- [ ] Security
- [ ] Test and Release
- [ ] User Experience
- [ ] Developer Infrastructure

**Please check any characteristics that apply to this pull request.**

- [ ] Does not have any [user-facing](https://github.com/istio/istio/tree/master/releasenotes#when-to-add-release-notes) changes. This may include CLI changes, API changes, behavior changes, performance improvements, etc.